requirements.txt
openai>=1.0.0
pinecone-client>=2.2.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.24.0
nltk>=3.8.0
transformers>=4.30.0
pydantic>=2.0.0
textblob>=0.17.1
pydub>=0.25.1 